Update with test

+++ ft_printf/print_s.c
@@ -6,7 +6,7 @@
 /*   By: chanhlee <chanhlee@student.42seoul.kr      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/21 23:02:57 by chanhlee          #+#    #+#             */
-/*   Updated: 2021/04/22 14:09:39 by chanhlee         ###   ########.fr       */
+/*   Updated: 2021/04/28 22:34:18 by chanhlee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int	print_string(char *str, t_opt *opt)
 		arg = ft_substr(str, 0, opt->prec);
 	else
 		arg = ft_strdup(str);
-	if (opt->width > ft_strlen(arg))
+	if (opt->width > (int)ft_strlen(arg))
 	{
 		padding = update_padding(opt->zero, opt->width - ft_strlen(arg));
 		arg = update_rest(arg, padding, opt->minus);
